Updated commander to accept a file of experiments to run.
Now, there's a '-f' option to commander that accepts a file that looks like this:
```
N:100
N:1000
TS:50,50,50
TS:100,100,100
TS:200,200,200
```

If an input file is passed, then the commander will test the tile sizes '50,50,50' for N=100 and N=1000.  Then it will test '100,100,100' for N=100 and N=1000.  And so on... for as many combos of N and TS (tile size) in the file. This way we can run arbitrary experiments in the future if need be.

The old recusive calls to search the tile space functionality is still there but it isn't performed if '-f' is used.

We can get commander to use a different TMM binary with the '--path-prefix' parameter. The code assumes that a file called 'TMM' exists in whatever path you pass (e.g, if '--path-prefix=./somedir' then TMM better exist in that directory -- ./somedir/TMM).

See 'experiment-config' for example.
See 'experiment-IJK-...sh" file for example of how to invoke.