Created 18 different loop/parallel versions to test.
To compile, run the './envVars.sh T', where T is the num of omp threads to test with.

Right now, the only correct one is the IJK version, the first and second loop folders.
Under the paralleltests/ijk/out/ directory there are the following folders:
* firstLoop
* secndLoop
* thirdLoop

Still need to figure out how to correctly parallelize the K-loops since multiple threads will be reading/writing the same accumulation spot.  For now, we can start gathering data for the parallelizing I and J loops.