Added K loop reductions.
These work, however they work by copying the entire R array many times and then adding the entire R array to itself multiple times. This means it is incredibly slow and should not ever really be used, but they do work. We measured a run time of 3 minutes on IJK-K with inputs of 5000 100 100 100.

This would be easier if TMM.c used block major format (we would not need to copy the whole thing).